Better Handle Multiple Models at Command Line
Previously, we were using ARGV to capture model names from the command
line. This worked until we started adding cli options, which got
vacuumed up with the other parts.

This commit changes to using the Thor .argument method to capture the
model names by position (as an array). We also move off of NamedBase and
onto Generator::Base, which is more semantically correct.

Fixes#16